apollo: was changed project: removed apollo-server; install apollo-server-express express; express wrapped in ApolloServer; was mounted intermediate app: server.applyMiddleware created routes
