+++ package.json
@@ -22,8 +22,10 @@
   },
   "homepage": "https://github.com/YanaLugina/photoshare-graphql#readme",
   "dependencies": {
-    "apollo-server": "^2.9.6",
+    "apollo-server-express": "^2.9.6",
+    "express": "^4.17.1",
     "graphql": "^14.5.8",
+    "graphql-playground-middleware-express": "^1.7.12",
     "nodemon": "^1.19.3"
   }
 }
