+++ typeDefs.graphql
@@ -1,0 +1,64 @@
+"""
+Date-time
+"""
+scalar DateTime
+
+"""
+Тип перечисления PhotoCategory
+"""
+enum PhotoCategory {
+    SELFIE
+    PORTRAIT
+    ACTION
+    LANDSCAPE
+    GRAPHIC
+}
+
+"""
+тип User
+"""
+type User {
+    githubLogin: ID!
+    name: String!
+    avatar: String
+    postedPhotos: [Photo!]!
+    inPhotos: [Photo!]!
+}
+
+"""
+тип Photo
+"""
+type Photo {
+    id: ID!
+    name: String!
+    url: String!
+    description: String
+    category: PhotoCategory!
+    postedBy: User!
+    taggedUsers: [User!]!
+    created: DateTime!
+}
+
+"""
+Тип ввода для мутации postPhoto
+"""
+input PostPhotoInput {
+    name: String!
+    description: String
+    category: PhotoCategory=PORTRAIT
+}
+
+"""
+Возвращаем Photo по запросу allPhotos
+"""
+type Query {
+    totalPhotos: Int!
+    allPhotos(after: DateTime): [Photo!]!
+}
+
+"""
+Возвращаем недавно опубликованную фотографию из мутации
+"""
+type Mutation {
+    postPhoto(input: PostPhotoInput): Photo!
+}
